Added Early Game Console model
Added in an early version of the Game Console model.

This model has animations for all of the buttons pressing down. The little triangle symbols with white lines coming out of the left of them are the animations and you can view it at the bottom right if you select it.