Added Horse Enemy to Model Test Scene
Added in a rigged early version of the enemy model with an early attack animation.

I placed the model into the Model Test scene of the test branch.

I also updated the Blender Files folder by adding in the updated blend files as of now, and placed the fbx files in them as well if anyone ever needs access to the fbx for other testing.