Added several new materials for objects in the main scene. Also added a new (but still unfinished) version of the room into the scene!
I added 2 more folders in the Models folder, the Bedroom folder, and the Decorations folder that is within the Bedroom folder. These folders have the materials that will hold the texture and normal maps of the models for the bedroom, keeping it organized for when the room is complete and is implemeted into the scene for good.

I also kept track of the scale size of the entire room model for when I import it to keep the size the same.

Position - X=2.25 Y=1.5 Z=-53
Rotation - X=0 Y=0 Z=0
Scale - X=0.8 Y=0.8 Z=0.8