Added in The finalized horse enemy model.
Added in the model for the Horse enemy that should be the final version. Added a lower set of teeth and new animations for standing idle, idle in the hallways, and for jumpscares.

The new model has been placed in the Models folder, within the Horse Enemy - Bottom Teeth folder.