Added in the Bedroom into the scene!
Added the entire model for the Bedroom into the main scene.

This added around 40 or so objects along with all of their textures (some very small details some larger things like room). I organized things as much as I could, especially in the inspector.

A portion of the models have materials that I had made manually because of them either having normal maps or me wanting to adjust the smoothness or metallic of the objects. Some of them still have the unchangeable material that is baked into the model. If you want to edit an objects material but it still has the unchangeable one, add a new material and put the image texture in the albedo slot.

You can find the prefab model for the bedroom and all of its textures/materials in the Bedroom folder that is within the Models folder .

PLEASE READ: - I had to add 2 separate models, the ceiling fan and the window because of some issues with textures. Not super important (wont effect anything) but wanted to give a reason why the fan and window were randomly separate.